python-ocr/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
easyocr==1.7.2
python-multipart==0.0.20
numpy==2.2.6
pillow==11.3.0
